タイトルシーンのBGM　リザルトシーン(浄化率ごとのBGM)を流す
