+++ CESA-2020-Prototype/Assets/Resources/Audio/Result76-100.mp3.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: f5b476de4ddb6764197b3ea9cb57aaed
+AudioImporter:
+  externalObjects: {}
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  preloadAudioData: 1
+  loadInBackground: 0
+  ambisonic: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
