+++ CESA-2020-Prototype/Assets/Scenes/TitleScene.unity
@@ -237,6 +237,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   select_frame_speed: 10
+  bgm: {fileID: 8300000, guid: 89962aba999918a47971ccb8b76f7cfd, type: 3}
 --- !u!4 &381011412
 Transform:
   m_ObjectHideFlags: 0
