--- CESA-2020-Prototype/Assets/Scripts/SoundFadeController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3972987d32f934c41ac4eba7b1673cf0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
